pyproject.toml
[project]
name = "wilsonrobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "pyyaml>=6.0.3",
    "scipy>=1.16.2",
]

